hub captureMessage
